+++ src/decision/decision_strategy.h
@@ -19,6 +19,7 @@
 #ifndef __CVC4__DECISION__DECISION_STRATEGY_H
 #define __CVC4__DECISION__DECISION_STRATEGY_H
 
+#include "preprocessing/assertion_pipeline.h"
 #include "prop/sat_solver_types.h"
 #include "smt/term_formula_removal.h"
 
@@ -57,9 +58,8 @@ public:
 
   bool needIteSkolemMap() override { return true; }
 
-  virtual void addAssertions(const std::vector<Node> &assertions,
-                             unsigned assertionsEnd,
-                             IteSkolemMap iteSkolemMap) = 0;
+  virtual void addAssertions(
+      const preprocessing::AssertionPipeline& assertions) = 0;
 };/* class ITEDecisionStrategy */
 
 class RelevancyStrategy : public ITEDecisionStrategy {
